Conflicts:
	cheatingday/src/main/java/com/icia/cheatingday/admin/service/rest/AdminRestService.java
	cheatingday/src/main/java/com/icia/cheatingday/manager/dao/StoreDao.java
	cheatingday/src/main/java/com/icia/cheatingday/review/dto/ReviewDto.java
	cheatingday/src/main/resources/mapper/orderMapper.xml
	cheatingday/src/main/resources/mapper/storeMapper.xml
	cheatingday/target/classes/com/icia/cheatingday/admin/controller/rest/AdminRestController.class
	cheatingday/target/classes/com/icia/cheatingday/admin/service/rest/AdminRestService.class
	cheatingday/target/classes/com/icia/cheatingday/manager/dao/StoreDao.class
	cheatingday/target/classes/com/icia/cheatingday/manager/service/MReviewOrderService.class
	cheatingday/target/classes/com/icia/cheatingday/order/dao/OrderDao.class
	cheatingday/target/classes/com/icia/cheatingday/review/dao/ReviewDao.class
	cheatingday/target/classes/com/icia/cheatingday/review/dto/ReviewDto$DtoForList$DtoForListBuilder.class
	cheatingday/target/classes/com/icia/cheatingday/review/dto/ReviewDto$DtoForList.class
	cheatingday/target/classes/com/icia/cheatingday/review/dto/ReviewDto$DtoForRead$DtoForReadBuilder.class
	cheatingday/target/classes/com/icia/cheatingday/review/dto/ReviewDto$DtoForRead.class
	cheatingday/target/classes/com/icia/cheatingday/review/dto/ReviewDto$DtoForUpdate.class
	cheatingday/target/classes/com/icia/cheatingday/review/dto/ReviewDto$DtoForWrite.class
	cheatingday/target/classes/com/icia/cheatingday/user/controller/mvc/UserController.class
	cheatingday/target/classes/com/icia/cheatingday/user/service/mvc/UserService.class
	cheatingday/target/classes/mapper/orderMapper.xml
	cheatingday/target/classes/mapper/storeMapper.xml

+++ cheatingday/src/main/java/com/icia/cheatingday/admin/dao/AdminDao.java
@@ -44,7 +44,7 @@ public class AdminDao {
 		return tpl.selectList("adminMapper.findAllByEnabled");
 	}
 	//[관리자]가입승인
-	public int enabledM(int mNum) {
+	public int enabledM(long mNum) {
 		return tpl.update("adminMapper.enabledM", mNum);
 	}
 	//[관리자]비상용 관리자가입
